src/main/java/com/rakovets/course/java/core/practice/date_and_time/DateWrapper.java
package com.rakovets.course.java.core.practice.date_and_time;

import java.time.LocalDate;
import java.time.Period;
import java.time.format.DateTimeFormatter;
import java.util.Locale;

public class DateWrapper {
    public static LocalDate getLocalDate(int year, int month, int day) {
        return LocalDate.of(year, month, day);
    }

    public static LocalDate getLocalDateWithChangedMonth(LocalDate date, int numberOfMonths) {
        return date.plusMonths(numberOfMonths);
    }

    public static String getStringWithDate(LocalDate date, String pattern) {
        return date.format(DateTimeFormatter.ofPattern(pattern, Locale.UK));
    }

    public static LocalDate makeLocalDateFromString(String date, String pattern) {
        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern(pattern, Locale.US);
        return LocalDate.parse(date, dateTimeFormatter);
    }

    public static int getNumberDayBetweenTwoDates(LocalDate date1, LocalDate date2) {
        Period period = Period.between(date1, date2);
        return period.getDays();
    }
}
